+++ cpp/tests/pagerank/pagerank_test.cu
@@ -123,9 +123,14 @@ class Tests_Pagerank : public ::testing::TestWithParam<Pagerank_Usecase> {
     ASSERT_EQ(fclose(fpin), 0);
 
     //  Pagerank runs on CSC, so feed COOtoCSR the row/col backwards.
-    cugraph::experimental::GraphCOOView<int, int, T> G_coo(&cooColInd[0], &cooRowInd[0], &cooVal[0], m, nnz);
+    cugraph::experimental::GraphCOOView<int, int, T> G_coo(
+      &cooColInd[0], &cooRowInd[0], &cooVal[0], m, nnz);
     auto G_unique = cugraph::coo_to_csr(G_coo);
-    cugraph::experimental::GraphCSCView<int, int, T> G(G_unique->view().offsets, G_unique->view().indices, G_unique->view().edge_data, G_unique->view().number_of_vertices, G_unique->view().number_of_edges);
+    cugraph::experimental::GraphCSCView<int, int, T> G(G_unique->view().offsets,
+                                                       G_unique->view().indices,
+                                                       G_unique->view().edge_data,
+                                                       G_unique->view().number_of_vertices,
+                                                       G_unique->view().number_of_edges);
 
     cudaDeviceSynchronize();
     if (PERF) {